SeapixGanApi/requirements.txt
annotated-types==0.7.0
anyio==3.6.2
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.4.0
click==8.0.4
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.1
fastapi==0.88.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
h11==0.13.0
idna==3.4
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
ipython-genutils==0.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
proglog==0.1.10
psutil==5.9.5
ptyprocess==0.7.0
pydantic==1.10.8
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-multipart==0.0.5
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.6
starlette==0.22.0
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.10.0
typing_extensions==4.8.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.19.0
zipp==3.20.2
